[#415] Display only one style of text fields in the demo app
