+++ demo/src/main/java/com/orange/ods/demo/ui/components/textfields/TextField.kt
@@ -19,21 +19,23 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import com.orange.ods.compose.component.textfield.OdsOutlinedTextField
+import com.orange.ods.compose.component.textfield.OdsTextField
+import com.orange.ods.compose.component.textfield.OdsTextFieldCharacterCounter
 import com.orange.ods.demo.R
+import com.orange.ods.demo.ui.components.textfields.TextFieldCustomizationState.Companion.TextFieldMaxChars
 import com.orange.ods.demo.ui.components.utilities.clickOnElement
 
 @Composable
-fun TextFieldOutlined(customizationState: TextFieldCustomizationState) {
+fun TextField(customizationState: TextFieldCustomizationState) {
     val context = LocalContext.current
     val trailingIconName = stringResource(id = R.string.component_element_trailing)
 
     Column {
-        OdsOutlinedTextField(
+        OdsTextField(
             modifier = Modifier
                 .fillMaxWidth()
                 .padding(top = dimensionResource(id = R.dimen.spacing_s)),
-            leadingIcon = if (customizationState.leadingIcon.value) painterResource(id = R.drawable.ic_heart) else null,
+            leadingIcon = if (customizationState.hasLeadingIcon) painterResource(id = R.drawable.ic_heart) else null,
             enabled = customizationState.isEnabled,
             isError = customizationState.isError,
             errorMessage = if (customizationState.isError) stringResource(id = R.string.component_text_field_error_message) else null,
@@ -56,3 +58,13 @@ fun TextFieldOutlined(customizationState: TextFieldCustomizationState) {
         )
     }
 }
+
+@Composable
+fun TextFieldCharacterCounter(valueLength: Int, enabled: Boolean) {
+    OdsTextFieldCharacterCounter(
+        valueLength = valueLength,
+        maxChars = TextFieldMaxChars,
+        enabled = enabled
+    )
+}
+
