+++ demo/src/main/res/values/strings.xml
@@ -245,8 +245,6 @@
     <string name="component_text_field">Text field</string>
     <string name="component_text_field_keyboard">Keyboard</string>
     <string name="component_text_fields_description">Text fields enable the user to type text into the app. They also provide ways to select text and the opportunity to save time by suggesting options for auto-completion.</string>
-    <string name="component_text_field_filled">Filled</string>
-    <string name="component_text_field_outlined">Outlined</string>
     <string name="component_text_field_text">Text field</string>
     <string name="component_text_field_password">Password field</string>
     <string name="component_text_field_input_type">Input type</string>
